+++ src/main/java/com/example/assignment/controller/PublishController.java
@@ -25,7 +25,7 @@ public class PublishController {
     private PostService postService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id ,
+    public String edit(@PathVariable(name = "id")Long id ,
                        Model model){
         PostDTO post= postService.getById(id);
         model.addAttribute("title", post.getTitle());
@@ -48,7 +48,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false)Integer id,
+            @RequestParam(value = "id", required = false)Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title", title);
